Added pseudo-synchronous message passing.
A pseudo-synchronous message will still be over UDP and has no real
failure detection. If a message is dropped, a message is dropped.
However, a client which makes a pseudo-synchronous request will be
able to set a timeout on the request. It will also be able to set
a function to handle a successful response to that request. This has
not been tested extensively and needs to be cleaned up. Also, the
code needs to be documented better. Therefore, I'm putting this in
a separate branch for the time being. Hopefully, I'll finish leader
election and reelection on leader failure, which will be a big step.